Fix Octomap's artifact installation strategy
Instead of setting the output for the build artifacts to be inside of the CMAKE_SOURCE_DIR, we use CMAKE_BINARY_DIR. For in-source builds, this will be the same as CMAKE_SOURCE_DIR, but for out-of-source builds this will install into the build/ folder (if that's what you named it).

I've also added an additional directory layer under bin/ and lib/. If you are using this as a component package in another CMake build, you might inadvertently copy all contents of the local bin and lib folders. I want to make sure we've isolated that to just octomap.

Fixes MarbleInc/mBot#885